refactor(cli): extract pma_tail_types.py from pma_thread_commands.py to isolate tail event dataclasses and simplify display_tail_events
